video-recommendation-engine/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
torch==2.0.0
torch-geometric==2.3.1
neo4j==5.8.1
pandas==2.0.1
numpy==1.24.3
requests==2.30.0
alembic==1.10.4
tqdm==4.65.0
scikit-learn==1.2.2
